refactor

+++ compute/src/main/java/com/tongji/DistributedSystem/bean/CallTimeBean.java
@@ -5,12 +5,12 @@ package com.tongji.DistributedSystem.bean;
  * All right reserved.
  */
 public class CallTimeBean {
-    public long timeSlot1 = 0,  // 0~3
-                timeSlot2 = 0,  // 3~6
-                timeSlot3 = 0,  // 6~9
-                timeSlot4 = 0,  // 9~12
-                timeSlot5 = 0,  // 12~15
-                timeSlot6 = 0,  // 15~18
-                timeSlot7 = 0,  // 18~21
-                timeSlot8 = 0;  // 21~24
+    public int  timeSlot1 = 0,  // 0~3
+            timeSlot2 = 0,  // 3~6
+            timeSlot3 = 0,  // 6~9
+            timeSlot4 = 0,  // 9~12
+            timeSlot5 = 0,  // 12~15
+            timeSlot6 = 0,  // 15~18
+            timeSlot7 = 0,  // 18~21
+            timeSlot8 = 0;  // 21~24
 }
